requirements.txt
asgiref==3.8.1
dj-rest-auth==7.0.1
Django==5.1.7
django-allauth==65.4.1
djangorestframework==3.15.2
sqlparse==0.5.3
